Refactor more (removed config)
