--- cuw.config.js
@@ -1,2 +1,0 @@
-// Custom Guess Who config file
-module.exports = {};
