--- src/App.css
@@ -1,22 +1,0 @@
-body {
-  font-family: 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
-  font-weight: 600;
-}
-
-.player1Color, .player1Color nameContainer {
-  background: rgb(0, 0, 117);
-  color: rgb(0, 0, 117);
-}
-
-.player2Color, .player2Color nameContainer {
-  background: #E10000;
-  color: #E10000;
-}
-
-.secretColor {
-  background: gold;
-}
-
-.newPage {
-  break-before: page;
-}
