--- src/constants/cardTypes.js
@@ -1,4 +1,0 @@
-export const CARD_TYPES = Object.freeze({
-  FACE_CARD: 1,
-  SECRET_CARD: 2
-});
