+++ src/constants/enums.js
@@ -1,0 +1,10 @@
+export const CARD_TYPES = Object.freeze({
+  FACE_CARD: 1,
+  SECRET_CARD: 2
+});
+
+export const CARD_COLOUR = {
+  PLAYER_1: "player1Color",
+  PLAYER_2: "player2Color",
+  SECRET: "secretColor"
+};
